Move `Shared` into its own file
