+++ tower/src/make/mod.rs
@@ -4,10 +4,11 @@ mod make_connection;
 mod make_service;
 
 pub use self::make_connection::MakeConnection;
-pub use self::make_service::{AsService, IntoService, MakeService, Shared};
+pub use self::make_service::shared::Shared;
+pub use self::make_service::{AsService, IntoService, MakeService};
 
 pub mod future {
     //! Future types
 
-    pub use super::make_service::SharedFuture;
+    pub use super::make_service::shared::SharedFuture;
 }
